+++ Framework/MainDriverApi.py
@@ -865,6 +865,7 @@ def run_test_case(
         file_specific_steps = all_file_specific_steps[TestCaseID] if TestCaseID in all_file_specific_steps else {}
         TestCaseName = testcase_info["title"]
         shared.Set_Shared_Variables("zeuz_current_tc", testcase_info, print_variable=False, pretty=False)
+        shared.Set_Shared_Variables("zeuz_auto_teardown", "on")
 
         # log_line = "# EXECUTING TEST CASE : %s :: %s #" % (test_case, TestCaseName)
         # print("#"*(len(log_line)))